+++ app/api/prog_image/helpers/api_response_helper.rb
@@ -1,0 +1,7 @@
+module ProgImage
+  module ApiResponseHelper
+    def say_ok
+      status :no_content
+    end
+  end
+end