+++ spec/api/prog_image/application_api_spec.rb
@@ -1,12 +1,10 @@
-RSpec.describe ApplicationApi, type: :request do
+RSpec.describe ProgImage::ApplicationApi, type: :request do
   describe 'GET /api/version' do
-    subject { decoded_json_response }
+    subject(:request) { decoded_json_body(dispatch(path: path)) }
 
     let(:path) { build_api_path(version: '', path: 'version') }
-    let(:version) { ProgImage::VERSION }
+    let(:version) { ProgImage.version }
     let(:latest_api_version) { configatron.api.latest_version.call }
-
-    before { dispatch path: path }
 
     its(['version']) { is_expected.to eq version }
     its(['latest_api_version']) { is_expected.to eq latest_api_version }