+++ spec/spec_helper.rb
@@ -1,11 +1,15 @@
+ENV['RACK_ENV'] ||= 'test'
+require File.expand_path('../config/application', __dir__)
+
+Dir[ProgImage.root.join('spec/support/**/*.rb')].each { |f| require f }
+
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
 
-  config.mock_with :rspec do |mocks|
-    mocks.verify_partial_doubles = true
-  end
+  config.include Rack::Test::Methods
+  config.include RequestSharedMethods, type: :request
 
   config.shared_context_metadata_behavior = :apply_to_host_groups
   config.filter_run_when_matching :focus