+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 #### Technical Improvements
 
 * Add Rubocop linting
+* Remove Rails and set-up as Grape only application